新增50张中国图片图集

+++ train.py
@@ -152,9 +152,10 @@ class YoloTrain(object):
         ckpt_list = []
         
         last_epoch = 0
+        last_test_loss = None
         if not from_scratch:
             try:
-                found = re.sub(r'.*yolov3_test_loss_(.*)\.ckpt-(\d+)', r'\1:\2', self.initial_weight)
+                found = re.sub(r'.*yolov3_800_test_loss_(.*)\.ckpt-(\d+)', r'\1:\2', self.initial_weight)
                 foundList = found.split(':')
                 last_test_loss, last_epoch = float(foundList[0]), int(foundList[1])
                 ckpt_list.append((self.initial_weight+" @", last_test_loss))
@@ -207,33 +208,36 @@ class YoloTrain(object):
                 test_epoch_loss.append(test_step_loss)
 
             # 记录该快照
-            limit = 5
+            limit = 6
             
             # 如果列表未满，且小于最大的损失率才添加
             train_epoch_loss, test_epoch_loss = np.mean(train_epoch_loss), np.mean(test_epoch_loss)
             
             if test_epoch_loss != None and math.isnan(float(test_epoch_loss)) == False:
-                # 开始存档
-                ckpt_file = "./checkpoint/yolov3_test_loss_%.4f.ckpt" % test_epoch_loss
-                log_start_time = time.localtime(time.time())
-                log_time = time.strftime('%Y-%m-%d %H:%M:%S', log_start_time)
-                print("=> Epoch: %d(%d/%d) Time: %s Train loss: %.2f Test loss: %.2f Saving %s ..."
-                                %(global_epoch, epoch, total_epoch, log_time, train_epoch_loss, test_epoch_loss, ckpt_file))
-                self.saver.save(self.sess, ckpt_file, global_step=global_epoch)
-                 
-                ckpt_list.append((ckpt_file+"-"+str(global_epoch), test_epoch_loss))
-                ckpt_list.sort(key=lambda elem: elem[1])
-                if len(ckpt_list) > limit:
-                    ckpt_list, remove_list = ckpt_list[:limit], ckpt_list[limit:]
-                    for elem in remove_list:
-                        fname = elem[0]    
-                        cmd = "rm "+fname+"*"
-                        print("calling ", cmd)
-                        os.system(cmd)
-                if len(ckpt_list) > 0:
-                    for k in range(len(ckpt_list)-1, -1, -1):
-                        elem = ckpt_list[k]
-                        print("current ranking:", elem[0])
+                if last_test_loss == None or test_epoch_loss < last_test_loss:
+                    # 开始存档
+                    ckpt_file = "./checkpoint/yolov3_800_test_loss_%.4f.ckpt" % test_epoch_loss
+                    log_start_time = time.localtime(time.time())
+                    log_time = time.strftime('%Y-%m-%d %H:%M:%S', log_start_time)
+                    print("=> Epoch: %d(%d/%d) Time: %s Train loss: %.2f Test loss: %.2f Saving %s ..."
+                                    %(global_epoch, epoch, total_epoch, log_time, train_epoch_loss, test_epoch_loss, ckpt_file))
+                    self.saver.save(self.sess, ckpt_file, global_step=global_epoch)
+                     
+                    ckpt_list.append((ckpt_file+"-"+str(global_epoch), test_epoch_loss))
+                    ckpt_list.sort(key=lambda elem: elem[1])
+                    if len(ckpt_list) > limit:
+                        ckpt_list, remove_list = ckpt_list[:limit], ckpt_list[limit:]
+                        for elem in remove_list:
+                            fname = elem[0]    
+                            loss = elem[1]
+                            if loss > last_test_loss:
+                                cmd = "rm "+fname+"*"
+                                print("calling ", cmd)
+                                os.system(cmd)
+                    if len(ckpt_list) > 0:
+                        for k in range(len(ckpt_list)-1, -1, -1):
+                            elem = ckpt_list[k]
+                            print("current ranking:", elem[0])
 
 
 if __name__ == '__main__': YoloTrain().train()
